RealTimeCrimeAnalytics/storm/requirements.txt
kafka-python==2.0.2
PyYAML==6.0.1
psycopg2-binary==2.9.9
pymongo==4.7.3
